+++ Framework/PWMapKit.framework/Versions/A/Headers/PWMapKit.h
@@ -2,54 +2,30 @@
 //  PWMapKit.h
 //  PWMapKit
 //
-//  Copyright (c) 2014 Phunware. All rights reserved.
+//  Created by Xiangwei Wang on 6/23/16.
+//  Copyright © 2016 Phunware. All rights reserved.
 //
 
 #import <Foundation/Foundation.h>
 
-#import "PWMapKitVersion.h"
-#import "PWMapView.h"
-#import "PWMapView+ZoomWorkaround.h"
-#import "PWSVPulsingAnnotationView.h"
-#import "PWUserTrackingBarButtonItem.h"
-#import "PWIndoorUserTracking.h"
-#import "PWAnnotationProtocol.h"
-#import "PWAnnotationType.h"
-#import "PWBuilding.h"
-#import "PWBuildingAnnotationProtocol.h"
-#import "PWBuildingAnnotationView.h"
-#import "PWBuildingAnnotation.h"
-#import "PWBuildingManager.h"
-#import "PWBuildingFloor.h"
-#import "PWBuildingFloorResource.h"
-#import "PWBuildingFloorReference.h"
-#import "PWBuildingOverlay.h"
-#import "PWBuildingOverlayRenderer.h"
-#import "PWMapDocument.h"
-#import "PWDirections.h"
-#import "PWDirectionsOptions.h"
-#import "PWDirectionsRequest.h"
-#import "PWDirectionsResponse.h"
-#import "PWRoute.h"
-#import "PWRouteStep.h"
-#import "PWAnnotationLabel.h"
-#import "PWIndoorUserLocation.h"
-#import "PWAccessibilityProtocol.h"
-#import "PWRouteManeuver.h"
-#import "PWRouteManeuverDirection.h"
-#import "PWRouteManeuverOverlay.h"
-#import "PWRouteManeuverOverlayRenderer.h"
-#import "PWMapKitAnalytics.h"
+#import <PWMapKit/PWMapView.h>
+#import <PWMapKit/PWBuilding.h>
+#import <PWMapKit/PWFloor.h>
+#import <PWMapKit/PWPointOfInterest.h>
+#import <PWMapKit/PWPointOfInterestType.h>
+#import <PWMapKit/PWCustomLocation.h>
+#import <PWMapKit/PWRoute.h>
+#import <PWMapKit/PWRouteInstruction.h>
 
-/**
- The `PWMapKit` framework provides an interface for embedding indoor maps directly into your own windows and views. This framework also provides support for annotating the map, finding your location, fetching routes, drawing routes and more.
- */
+/** Light weight UI **/
+#import <PWMapKit/PWMapViewController.h>
+#import <PWMapKit/PWBuildingViewController.h>
+#import <PWMapKit/PWRouteViewController.h>
+#import <PWMapKit/PWRouteInstructionsView.h>
+#import <PWMapKit/PWRouteInstructionsViewController.h>
+#import <PWMapKit/PWPointOfInterestViewController.h>
 
 @interface PWMapKit : NSObject
-
-///----------------------
-/// @name Utility Methods
-///----------------------
 
 /**
  Returns PWMapKit.