+++ Framework/PWMapKit.framework/Versions/A/Headers/PWRoute.h
@@ -2,39 +2,68 @@
 //  PWRoute.h
 //  PWMapKit
 //
-//  Copyright (c) 2015 Phunware. All rights reserved.
+//  Created by Steven Spry on 5/12/16.
+//  Copyright © 2016 Phunware. All rights reserved.
 //
 
-#import "PWMappingTypes.h"
+#import <Foundation/Foundation.h>
+
+#import "PWBuilding.h"
+#import "PWPointOfInterest.h"
 
 /**
- The `PWRoute` class defines a single route that the user can follow between a requested start and end point. The route object defines the geometry for the route and includes route information you can display to the user, such as the name of the route, its distance and the expected travel time.
- 
- Do not create instances of this class directly. Instead, request directions to receive route objects. For more information about requesting directions, see `PWDirections` Class Reference.
+ *  A PWRoute represents a route calculated for the user from within a building's boundaries start point of interest to end point of interest.
  */
-
 @interface PWRoute : NSObject
 
-/**
- The route distance in meters. (read-only)
- @discussion This property reflects the distance the user traverses on the route's path. It is not a direct distance between the start and end points of the route.
+/**---------------------------------------------------------------------------------------
+ * @name Properties
+ *  ---------------------------------------------------------------------------------------
  */
-@property (readonly) CLLocationDistance distance;
 
 /**
- The array of steps that comprise the overall route. (read-only)
- @discussion The array contains one or more `PWRouteStep` objects representing distinct portions of the route. Each step corresponds to a single floor sequence that must be followed along the route.
+ *  A reference to the route's building object.
  */
-@property (readonly) NSArray *steps;
+@property (nonatomic,readonly) PWBuilding *building;
 
 /**
- The array of maneuvers that a user is expected to perform in the traversal of this route.
+ *  A reference to the origin, or start point of interest, for the route.
  */
-@property (readonly) NSArray /* PWRouteManeuver */ *maneuvers;
+@property (nonatomic,readonly) PWPointOfInterest *startPointOfInterest;
 
 /**
- A Boolean value that indicates whether the PWRoute object is accessible. (read-only)
+ *  A reference to the destination, or end point of interest, for the route.
  */
-@property (readonly, getter=isAccessible) BOOL accessible;
+@property (nonatomic,readonly) PWPointOfInterest *endPointOfInterest;
+
+/**
+ *  An array of PWRouteInstruction objects containing instructions to follow the route path.
+ */
+@property (nonatomic,readonly) NSMutableArray *routeInstructions;
+
+/**
+ *  An integer number representing the total distance of the route expressed in meters.
+ */
+@property (nonatomic,readonly) NSInteger distance;
+
+/**
+ *  An integer number representing the estimated time of the route expressed in seconds.
+ */
+@property (nonatomic,readonly) NSInteger estimatedTime;
+
+/**---------------------------------------------------------------------------------------
+ * @name Class Methods
+ *  ---------------------------------------------------------------------------------------
+ */
+
+/**
+ *  Instantiates a new PWRoute object using the given parameters. Calculates a route to navigate between the start POI and the end POI. The completion handler is called when the route is fully calculated.
+ *
+ *  @param startPoint    PWPointOfInterest object representing the start point for the route calculation.
+ *  @param endPoint      PWPointOfInterest object representing the end point for the route calculation.
+ *  @param accessibility BOOL Value to tell the route init if accessibility should be considered to calculate the route.
+ *  @param completion    Completion handler that is called once the routes calculation is complete.
+ */
++ (void)initRouteFrom:(PWPointOfInterest *)startPoint to:(PWPointOfInterest *)endPoint accessibility:(BOOL)accessibility completion:(void(^)(PWRoute *route, NSError *error))completion;
 
 @end