+++ Samples/LoadMap/LoadMap/AppDelegate.h
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.h
+//  LoadMap
+//
+//  Created by Xiangwei Wang on 7/25/16.
+//  Copyright © 2016 Phunware Inc. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+
+@end
+